Compagins Finsh

+++ src/Core/EndProject.Application/DTOs/Car/CarCampaignsDTO.cs
@@ -5,6 +5,6 @@ public class CarCampaignsDTO
     public bool isCampaigns { get; set; } = true;
     public DateTime PickUpCampaigns { get; set; }
     public DateTime ReturnCampaigns { get; set; }
-    public double CampaignsInterest { get; set; }
+    public decimal CampaignsInterest { get; set; }
     public string SuperAdminId { get; set; }
 }
